pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gen-ai-utils"
version = "1.0.0"
description = "Comprehensive utilities for AI/ML, Data Science, and Data Engineering"
authors = [
    {name = "Gen AI Utils Team"}
]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
keywords = ["ai", "ml", "data-science", "data-engineering", "openai", "utilities"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "openai>=1.12.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "scikit-learn>=1.4.0",
    "requests>=2.31.0",
    "colorama>=0.4.6",
    "pyyaml>=6.0.0",
    "tqdm>=4.66.0",
    "pytz>=2024.1",
]

[project.optional-dependencies]
llamaparse = [
    "llama-parse>=0.4.0",
    "llama-index>=0.10.0",
]
streamlit = [
    "streamlit>=1.31.0",
]
vectordb = [
    "chromadb>=0.4.22",
]
advanced-ml = [
    "umap-learn>=0.5.5",
    "statsmodels>=0.14.0",
]
media = [
    "youtube-transcript-api>=0.6.2",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.1.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "ipython>=8.21.0",
]
all = [
    "gen-ai-utils[llamaparse,streamlit,vectordb,advanced-ml,media,dev]",
]

[project.urls]
Homepage = "https://github.com/yourusername/gen-ai-utils"
Repository = "https://github.com/yourusername/gen-ai-utils"
Documentation = "https://github.com/yourusername/gen-ai-utils#readme"

[tool.setuptools.packages.find]
where = ["."]
include = ["gen_ai_utils*"]
exclude = ["tests*", "docs*", "examples*"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=gen_ai_utils --cov-report=html --cov-report=term"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
